Move player gesture handling to extra class
